refactor(feeds): use Int sortOrder for greader sortid
- Revert feed `sortOrder` type from `Long` to `Int`.
- Parse `sortid` with `toIntOrNull(16)`.
- Align sorting fallback and tests with Int-based sort order.

+++ app/src/main/java/me/ash/reader/domain/data/FeedSorting.kt
@@ -9,7 +9,7 @@ fun sortFeedsAlphabetically(feeds: List<Feed>): List<Feed> {
 
 fun sortFeedsBySortOrder(feeds: List<Feed>): List<Feed> {
     return feeds.sortedWith(
-        compareBy<Feed> { it.sortOrder ?: Long.MAX_VALUE }
+        compareBy<Feed> { it.sortOrder ?: Int.MAX_VALUE }
             .thenBy(String.CASE_INSENSITIVE_ORDER) { it.name }
     )
 }
